refactor(extension): 修改全局状态，剥离通信相关代码
