+++ packages/extension/src/types/config.ts
@@ -1,6 +1,8 @@
+export type ProviderType = 'ollama' | 'openai' | 'openrouter';
+
 export interface ChatModel {
     id: string;
-    type: 'openai' | 'ollama' | 'openrouter';
+    type: ProviderType;
     model: string;
     title?: string;
     baseURL?: string;
