+++ packages/gui-chat/src/api/vscode.ts
@@ -1,0 +1,5 @@
+declare const acquireVsCodeApi: () => {
+    postMessage: (message: any) => undefined;
+};
+
+export const vscode = (typeof acquireVsCodeApi === 'function') ? acquireVsCodeApi() : undefined;
