+++ packages/gui-chat/src/components/popup/DeleteModel.vue
@@ -16,7 +16,7 @@
 
 <script setup lang="ts">
 import type { Model } from '@/types'
-import { useSenderStore } from '@/stores/sender'
+import { modelDelete } from '@/api/sender';
 
 const props = defineProps<{
   deleteModel: Model | undefined,
@@ -25,7 +25,7 @@ const props = defineProps<{
 
 function confirm() {
   const modelID = props.deleteModel?.id || '';
-  useSenderStore().modelDelete(modelID)
+  modelDelete(modelID)
   props.popupDeleteModel()
 }
