+++ packages/gui-chat/src/i18n/index.ts
@@ -4,13 +4,15 @@ import zh_cn from './lang/zh-cn';
 import en from './lang/en';
 import ja from './lang/ja';
 
+export type Language = 'en' | 'zh-cn' | 'ja';
+
 const i18n = createI18n({
     legacy: false,
     locale: 'en',
     messages: {
-        zh_cn,
-        en,
-        ja
+        "zh-cn": zh_cn,
+        "en": en,
+        "ja": ja,
     }
 });
