+++ packages/gui-chat/src/types/config.ts
