+++ packages/gui-chat/src/types/dialog.ts
