+++ packages/gui-chat/src/types/index.ts
@@ -1,2 +1,2 @@
 export * from './dialog'
-export * from './model'
+export * from './config'
